Breakout styles, improve animation
