+++ package.json
@@ -37,6 +37,7 @@
     "react-hook-form": "^6.3.0",
     "react-modal": "^3.16.1",
     "react-responsive": "^9.0.2",
+    "react-range-slider-input": "^3.0.7",
     "react-scripts": "4.0.3",
     "react-spinners": "^0.11.0",
     "react-styleguidist": "^11.1.7",
@@ -85,7 +86,6 @@
     "jest": "^26.1.0",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.8.8",
-    "react-range-slider-input": "^3.0.7",
     "semantic-release": "^17.4.6"
   }
 }
